requirements.txt
appdirs==1.4.4
arrow==0.17.0
asgiref==3.3.0
attrs==20.3.0
black==20.8b1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
coverage==5.3
cryptography==3.2.1
defusedxml==0.6.0
Django==3.0
django-allauth==0.43.0
django-audit-fields==0.1.15
django-crypto-fields==0.3.0
django-currentuser==0.5.2
django-debug-toolbar==3.1.1
django-extensions==3.1.1
django-preventconcurrentlogins==0.8.2
django-revision==0.1.20
django-simple-history==2.12.0
docopt==0.6.2
edc-constants==0.3.1
edc-model-fields==0.3.2
edc-utils==0.1.24
factory-boy==3.1.0
Faker==0.9.1
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
iniconfig==1.1.1
isort==5.7.0
Markdown==3.2.1
mixer==6.1.3
model-bakery==1.2.1
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.4
pathspec==0.8.1
pipreqs==0.4.10
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pycryptodome==3.10.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
python-dateutil==2.8.1
python-environ==0.4.54
python3-openid==3.2.0
pytz==2020.1
qrcode==6.1
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
smmap==3.0.4
sqlparse==0.4.1
text-unidecode==1.2
toml==0.10.2
tqdm @ file:///tmp/build/80754af9/tqdm_1609788246169/work
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.11
wifi-password==1.0.7
yarg==0.1.9
